requirements.txt
tensorflow==2.13.0
pyqt5==5.15.9
numpy==1.22.0
Image==1.5.33
opencv-python==4.8.0.76
pyfiglet==0.8.post1
mss==9.0.1
pillow==10.0.0
black==23.7.0
pytest==7.4.0